refactor: デッドコード削除・重複定数共通化・theme-toggle移動
- monthly-overview/index.tsxから未使用の計算ロジックとrecentSummaries propを削除
- typeLabelsの重複定義をsrc/lib/constants.tsに共通化
- theme-toggle.tsxをsrc/components/features/からsrc/components/ui/に移動
