+++ src/lib/constants.ts
@@ -1,0 +1,5 @@
+export const TYPE_LABELS = {
+  income: '収入',
+  expense: '支出',
+  carryover: '繰越',
+} as const
