+++ tests/components/features/theme-toggle.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { ThemeToggle } from '@/components/features/theme-toggle'
+import { ThemeToggle } from '@/components/ui/theme-toggle'
 
 const mockSetTheme = vi.fn()
